Revert "try to fix ubigint percentile range"
This reverts commit 9d028e8de29969d8991e699b63e4d8b0f15624f2.

+++ src/query/src/qPercentile.c
@@ -74,7 +74,7 @@ static int32_t setBoundingBox(MinMaxEntry* range, int16_t type, double minval, d
     }
   } else if (IS_UNSIGNED_NUMERIC_TYPE(type)){
     range->u64MinVal = (uint64_t) minval;
-    if (maxval > UINT64_MAX) {
+    if ((uint64_t)maxval > UINT64_MAX) {
       range->u64MaxVal = UINT64_MAX;
     } else {
       range->u64MaxVal = (uint64_t) maxval;
